WC-766/refactor:파일 구조 변경 및 복약관리 레이아웃 제작을 위한 리펙토링
