+++ wanchcoach/src/app/(route)/mainpage/medication/medication.module.css
@@ -1,0 +1,13 @@
+.container {
+  background-color: #f9f9f9;
+  height: calc(100vh - 60px - 70px);
+  width: 100%;
+}
+
+.header_menu_container {
+  display: flex;
+  width: max-content;
+  margin: 0 8%;
+  height: 12vh;
+  align-items: center;
+}
