+++ wanchcoach/src/app/(route)/mainpage/treatment/_components/component.tsx
