+++ wanchcoach/src/app/(route)/mainpage/treatment/diagnosis/diagnosis.module.css
