+++ wanchcoach/src/app/(route)/mainpage/treatment/diagnosis/layout.tsx
