+++ wanchcoach/src/app/(route)/mainpage/treatment/treatment.module.css
@@ -4,28 +4,6 @@
   width: 100%;
 }
 
-.header_profile_container {
-  text-align: -webkit-right;
-  margin: 0 8%;
-  height: 60px;
-  align-content: center;
-}
-.header_plus_box {
-  display: flex;
-  align-items: center;
-  height: max-content;
-  width: max-content;
-  padding: 6px 15px;
-  border-radius: 20px;
-  box-shadow: 2px 4px 4px #dddddd, -0.3px -0.3px 1px #dddddd;
-  cursor: pointer;
-}
-.header_plus_text {
-  font-size: 16px;
-  font-weight: 600;
-  margin-left: 6px;
-}
-
 .header_menu_container {
   display: flex;
   width: max-content;
@@ -34,57 +12,6 @@
   align-items: center;
 }
 
-.header_menu_box {
-  border-radius: 10px;
-  background-color: "white";
-  width: 85px;
-  height: max-content;
-  text-align: -webkit-center;
-  margin: 10px 2%;
-  padding: 1vh 0;
-}
-
-.header_menu_text {
-  font-size: 15px;
-  font-weight: 600;
-  margin-top: 2px;
-}
-
-.modal_header {
-  display: flex;
-  height: 8vh;
-  justify-content: center;
-  align-items: center;
-}
-.modal_header_text_01 {
-  font-size: 20px;
-  font-weight: 600;
-  padding: 0 5px;
-  padding-right: 10px;
-}
-.modal_header_text_02 {
-  font-size: 14px;
-  font-weight: 600;
-  padding: 0 5px;
-  margin-top: 5px;
-}
-.modal_content {
-  height: 14vh;
-  display: flex;
-  justify-content: space-around;
-  padding: 10px;
-}
-.modal_content_box {
-  width: 40%;
-  border-radius: 10px;
-  align-content: center;
-  text-align: -webkit-center;
-}
-.modal_content_text {
-  font-size: 14px;
-  padding-top: 10px;
-  font-weight: 600;
-}
 /* component.tsx */
 
 /* TreatmentBox() */
