+++ wanchcoach/src/app/(route)/register/page.tsx
