--- wanchcoach/src/app/mainpage/medication/layout.tsx
@@ -1,5 +1,0 @@
-import { ReactNode } from "react";
-
-export default function MedicationLayout({ children }: { children: ReactNode }) {
-  return <>{children}</>;
-}
